Update windows-in-place-upgrade.md
While working with the customer on a case, where he had migrated the VM from On-Prem to Azure. When OS was on the VM was installed they used German Language and later migrated to Azure.
Now they wanted to perform an in-place upgrade of the VM which was not allowing to "Keep Files and Folders" instead was selecting "Nothing"(Clean Installation).

Media is only available in English [US] volume license. No other languages or editions are supported. This is not mentioned in any Public Article for the customer to refer and proceed.