Update resize-vm.md
Added a note section about not seeing the SKU in the Portal if it does/does not have a temp disk. Also fixed some spacing on the bullets under the portal section and adjusted the line about the workaround for going from no temp disk to a temp disk VM (even though I feel like that should be in the linked documentation).