Update dasv4-series.md
The original size name was listed as Standard_D2as_v42, but the correct name is Standard_D2as_v4. A number 2 was mistakenly added at the end of the name. I checked the Azure Portal to validate whether the information provided here matches what is shown there. The information matched, meaning that the VM size listed as Standard_D2as_v42 corresponds to Standard_D2as_v4 in the Azure Portal. (In Azure Portal the name listed is Standard_D2as_v4 and the limits observed match with Standard_D2as_v42 in public doc.)