--image-version is incorrect parameter
The command az sig image-version create uses the --image-version parameter with a managed image’s resource ID, which is incorrect. The --image-version flag is meant for specifying an existing image version (in a gallery) as the source, whereas a managed image source requires the --managed-image parameter. 

Also, the az sig image-version create will fail if there is not already an image definition present.  You may want to add a step to make sure and create the image definition using az sig image-definition create